requirements.txt
fastapi==0.111.0
pymemcache==4.0.0
pyodbc==5.1.0
python-dotenv==1.0.1
redis==5.0.6
uvicorn==0.30.1
